# OAS/openapi.yaml
openapi: "3.0.0"
servers:
  - url: /ori-mock
info:
  title: "De ORI API voor Open Overheid"
  description: "Gegevens uit de Open Raads Informatie systemen kunnen met deze API worden aangeleverd voor actieve openbaarmaking via Open Overheid. Ook kan deze API gebruikt worden om aangeleverde data bij te werken. Deze versie van deze API zal naast documentobjecten ook vergader-informatie kunnen verwerken volgens het ontwerp van de VNG mini-ORI API, en daarbij ook een endpoint voor Vergaderingen en Agendapunten ontsluiten. Bij gebruik door decentrale overheden zal de DigiKoppeling-standaard verplicht worden voor het aanleveren bij Open Overheid (zie Forum Standaardisatie). Nu wordt nog de Client Credentials flow toegepast. ([OAuth 2.0](https://swagger.io/docs/specification/authentication/oauth2/)) Deze specificatie wordt met een uitgebreide toelichting onderhouden op https://gitlab.com/koop/woo/aanleveren-ori "
  version: "0.1.0"
  x-imvertor-generator-version: "2.0.0"
  x-yamlCompiler-stylesheets-version: "20230315"
  contact:
    url: "https://gitlab.com/koop/woo/aanleveren-ori"
  license:
    name: "European Union Public License, version 1.2 (EUPL-1.2)"
    url: "https://eupl.eu/1.2/nl/"
paths:
...
  /informatieobjecten:
    get:
      operationId: "getinformatieobjecten"
      summary: "Alle informatieobjecten opvragen."
      description: "Deze lijst kan in de toekomst gefilterd worden met behulp van query parameters."
      responses:
        200:
          description: "Zoekactie geslaagd"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInformatieObjectList"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Informatieobjecten"
    post:
      operationId: "postinformatieobject"
      description: "Het vastleggen van een informatieobject"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InformatieObjectZonderPid"
      responses:
        201:
          description: "Succesvol aangeleverd."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InformatieObject"
        400:
          description: "Aanleverfout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
      tags:
        - "Informatieobjecten"
  /informatieobjecten/{id}:
    delete:
      operationId: "delinformatieobject"
      description: "Het bericht voor het verwijderen van een informatieobject."
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
            title: ID
          required: true
        - in: header
          name: X-Reason
          schema:
            type: string
            description: "Reden waarom een informatieobject wordt verwijderd."
          required: true
      responses:
        200:
          description: "Verwijderactie geslaagd"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Informatieobjecten"
    get:
      operationId: "getinformatieobject"
      summary: "Een specifiek informatieobject opvragen."
      description: "Een specifiek informatieobject opvragen op basis van een identifier."
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
            title: ID
          required: true
      responses:
        200:
          description: "Zoekactie geslaagd"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InformatieObject"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Informatieobjecten"
    put:
      operationId: "putinformatieobject"
      description: "Het wijzigen van een Informatieobject"
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
            title: ID
          required: true
        - in: header
          name: X-Reason
          schema:
            type: string
            description: "Reden waarom een informatieobject wordt gewijzigd."
          required: true
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - properties:
                    organisatie:
                      readOnly: true
                - $ref: "#/components/schemas/InformatieObject"
      responses:
        201:
          description: "Succesvol gewijzigd."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InformatieObject"
        400:
          description: "Aanleverfout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: "Deze wijziging bevat aanpassingen op velden die niet mogen worden aangepast."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Informatieobjecten"
...
components:
  schemas:
...
    InformatieObject:
      type: "object"
      description: "Een (document-)object dat tijdens een vergadering is besproken of behandeld"
      required:
        - "pid"
        - "webpaginalink"
        - "organisatie"
        - "titel"
        - "wooinformatiecategorie"
        - "datumingediend"
      properties:
        pid:
          type: string
          format: uri
          readOnly: true
          title: url
          description: "Persistent Identifier, URL-referentie naar dit object binnen Open Overheid. Dit is de unieke identificatie en locatie van dit object."
          minLength: 1
        webpaginalink:
          title: Webpagina link
          description: "De URL naar de webpagina waar dit document bij het bestuursorgaan is terug te vinden."
          type: "string"
          format: "uri"
          example: "https://example.com/mijn-pagina.html"
        organisatie:
          $ref: "#/components/schemas/Organisatie"
        titel:
          type: "string"
          title: "Titel"
          description: "De titel van het informatieobject."
        wooinformatiecategorie:
          type: string
          title: Woo Informatiecategorie
          description: "De code van de informatiecategorie volgens TOOI. Bijvoorbeeld de code 'c_db4862c3' voor 'vergaderstuk decentrale overheid'."
          example: "c_db4862c3"
        datumingediend:
          type: "string"
          title: "Datum ingediend"
          description: "De datum waarop het informatieobject als vergaderstuk is ingediend."
          format: "date"
          example: "2017-02-09"
        id:
          type: string
          description: "Unieke resource identifier."
          title: ID
        auteur:
          type: "string"
          title: "Auteur"
          description: "De naam van de auteur van het document."
        bronorganisatie:
          type: "string"
          title: "Bronorganisatie"
          description: "De organisatie die verantwoordelijk is voor de inhoud van een document. Deze kan met een TOOI-uri worden aangeduid maar het kan ook een niet-publieke organisatie betreffen."
        creatiedatum:
          type: "string"
          title: "Creatiedatum"
          description: "Creatiedatum. Moment waarop het document is gemaakt."
        informatieobjecttype:
          $ref: "#/components/schemas/Informatieobjecttype"
        formaat:
          type: string
          title: Formaat
          description: "Het Media Type (voorheen MIME type) voor de wijze waarop de inhoud van het informatieobject is vastgelegd in een computerbestand. Voorbeeld: `application/pdf`."
        omschrijving:
          type: "string"
          title: "Omschrijving"
          description: "De omschrijving van het informatieobject."
        taal:
          type: "string"
          title: "Taal"
          description: "De taal van het informatieobject. Bij voorkeur aaangeduid volgens ISO 639-2/B."
        vergaderingen:
          type: array
          items:
            $ref: "#/components/schemas/VerwijzingNaarResource"
          description: "Verwijzingen naar de vergaderingen die dit informatieobject als bijlagen hebben."
          uniqueItems: true
        agendapunten:
          type: "array"
          title: "Agendapunten"
          description: "Verwijzingen naar de agendapunten die dit informatieobject als bijlagen hebben."
          items:
            $ref: "#/components/schemas/VerwijzingNaarResource"
          uniqueItems: true
        gerelateerdinformatieobject:
          $ref: "#/components/schemas/GerelateerdInformatieobject"
    InformatieObjectZonderPid:
      type: "object"
      description: "Een (document-)object dat tijdens een vergadering is besproken of behandeld. Bij aanlevering krijgt een object een Pid van Open Overheid."
      required:
        - "webpaginalink"
        - "organisatie"
        - "titel"
        - "wooinformatiecategorie"
        - "datumingediend"
      properties:
        webpaginalink:
          title: Webpagina link
          description: "De URL naar de webpagina waar dit document bij het bestuursorgaan is terug te vinden."
          type: "string"
          format: "uri"
          example: "https://example.com/mijn-pagina.html"
        organisatie:
          $ref: "#/components/schemas/Organisatie"
        titel:
          type: "string"
          title: "Titel"
          description: "De titel van het informatieobject."
        wooinformatiecategorie:
          type: string
          title: Woo Informatiecategorie
          description: "De code van de informatiecategorie volgens TOOI. Bijvoorbeeld de code 'c_db4862c3' voor 'vergaderstuk decentrale overheid'."
          example: "c_db4862c3"
        datumingediend:
          type: "string"
          title: "Datum ingediend"
          description: "De datum waarop het informatieobject als vergaderstuk is ingediend."
          format: "date"
          example: "2017-02-09"
        id:
          type: string
          description: "Kenmerk van het object binnen het bestuursorgaan zelf."
          title: ID
        auteur:
          type: "string"
          title: "Auteur"
          description: "De naam van de auteur van het document."
        bronorganisatie:
          type: "string"
          title: "Bronorganisatie"
          description: "De organisatie die verantwoordelijk is voor de inhoud van een document. Deze kan met een TOOI-uri worden aangeduid maar het kan ook een niet-publieke organisatie betreffen."
        creatiedatum:
          type: "string"
          title: "Creatiedatum"
          description: "Creatiedatum. Moment waarop het document is gemaakt."
        informatieobjecttype:
          $ref: "#/components/schemas/Informatieobjecttype"
        formaat:
          type: string
          title: Formaat
          description: "Het Media Type (voorheen MIME type) voor de wijze waarop de inhoud van het informatieobject is vastgelegd in een computerbestand. Voorbeeld: `application/pdf`."
        omschrijving:
          type: "string"
          title: "Omschrijving"
          description: "De omschrijving van het informatieobject."
        taal:
          type: "string"
          title: "Taal"
          description: "De taal van het informatieobject. Bij voorkeur aaangeduid volgens ISO 639-2/B."
        vergaderingen:
          type: array
          items:
            $ref: "#/components/schemas/VerwijzingNaarResource"
          description: "Verwijzingen naar de vergaderingen die dit informatieobject als bijlagen hebben."
          uniqueItems: true
        agendapunten:
          type: "array"
          title: "Agendapunten"
          description: "Verwijzingen naar de agendapunten die dit informatieobject als bijlagen hebben."
          items:
            $ref: "#/components/schemas/VerwijzingNaarResource"
          uniqueItems: true
        gerelateerdinformatieobject:
          $ref: "#/components/schemas/GerelateerdInformatieobject"
    GerelateerdInformatieobject:
      type: object
      title: Gerelateerd informatieobject
      description: "Relevant ander informatieobject dat gerelateerd is aan dit informatieobject."
      nullable: true
      required:
        - informatieobject
        - rol
      properties:
        informatieobject:
          type: "string"
          description: "De identificatie van het gerelateerde document."
        rol:
          type: string
          title: Rol
          description: "Het type van de relatie. bijlage_van: https://identifier.overheid.nl/tooi/def/thes/kern/c_4d1ea9ba "
...
    Informatieobjecttype:
      type: "string"
      description: "Het type van het informatieobject wordt gebruikt om zoekresultaten te kunnen filteren en gebruikers snel inzicht te geven in de aard van het document."
...
    VerwijzingNaarResource:
      type: object
      required:
        - url
        - id
      properties:
        id:
          description: "ID van de resource waarnaar verwezen wordt."
          type: string
          minLength: 1
        url:
          description: "URL-referentie naar de resource waarnaar verwezen wordt."
          type: string
          format: uri
          minLength: 1
          readOnly: true
    Gemeente:
      type: object
      properties:
        gemeente:
          type: string
          title: Gemeente
          description: 'De code van de gemeente volgens TOOI.
            Bijvoorbeeld "gm0363" voor gemeente Amsterdam.'
          example: gm0363
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Gemeente Amsterdam"
      required:
        - gemeente
        - naam
    Provincie:
      type: object
      properties:
        provincie:
          type: string
          title: Provincie
          description: 'De code van de provincie volgens TOOI.
            Bijvoorbeeld "pv27" voor provincie Noord-Holland.'
          example: pv27
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Provincie Groningen"
      required:
        - provincie
        - naam
    Waterschap:
      type: object
      properties:
        waterschap:
          type: string
          title: Waterschap
          description: 'De code van het waterschap volgens TOOI.
            Bijvoorbeeld "ws0654" voor het waterschap Aa en Maas.'
          example: ws0654
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Waterschap Aa en Maas"
      required:
        - waterschap
        - naam
    Organisatie:
      oneOf:
        - $ref: "#/components/schemas/Gemeente"
        - $ref: "#/components/schemas/Provincie"
        - $ref: "#/components/schemas/Waterschap"
...
    PaginatedInformatieObjectList:
      type: object
      properties:
        next:
          type: string
          format: uri
          nullable: true
          description: "URL naar de volgende pagina met resultaten."
        previous:
          type: string
          format: uri
          nullable: true
          description: "URL naar de vorige pagina met resultaten."
        results:
          type: array
          items:
            $ref: "#/components/schemas/InformatieObject"
...
    ErrorResponse:
      type: object
      properties:
        titel:
          type: string
          description: "Titel van de foutmelding."
        status:
          type: integer
          format: int32
          description: "De HTTP statuscode van het antwoord."
        detail:
          type: string
          description: "Omschrijving van de foutmelding."